magit-completing-read: Improve handling of default
Consistently put it at the front of the completion candidates, while
making sure it isn't repeated later in the collection, taking into
account that some or all entries of COLLECTION may be atoms or
cons-cell, and that DEF may also be a cons-cell.  Also if DEF is a
cons-cell, then make sure to only use its car in the prompt and when
returning it without performing any completion.